% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aupdc.R
\name{aupdc}
\alias{aupdc}
\title{AUPDC}
\usage{
aupdc(y, time)
}
\arguments{
\item{y}{Ammount of the disease.}

\item{time}{Measurement time}
}
\value{
area Area under progress disease curve
}
\description{
AUPDC
}
\examples{
day <- seq(10,60,10)
sev <- c(0.22, 0.26, 0.28, 0.36, 0.39, 0.42)
aacpd(y=sev, time=day)
}
