% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expon.R
\name{expon}
\alias{expon}
\title{Fit Exponential model}
\usage{
expon(y, time)
}
\arguments{
\item{y}{Proportion of the disease. Must be a value between 0 and 1.}

\item{time}{Measurement time}
}
\value{
r Growth rate

y0 Initial inoculum

time Measurement time

pred Predicted data

r2 Determination coefficient between the observed and predicted data
}
\description{
Fit Exponential model
}
\examples{
dap <- seq(1, 70, 10)
inc <- c(0.01, 0.04, 0.05, 0.11, 0.30, 0.55, 0.94)
expon(y=inc, time=dap)
}
