% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.R
\name{lin}
\alias{lin}
\title{Fit linear model}
\usage{
lin(y, time)
}
\arguments{
\item{y}{Proportion of the disease. Must be a value between 0 and 1.}

\item{time}{Measurement time}
}
\value{
r Growth rate

y0 Initial inoculum

time Measurement time

pred Predicted data

r2 Determination coefficient between the observed and predicted data
}
\description{
Fit linear model
}
\examples{
dap <- seq(10, 60, 10)
inc <- c(0.18, 0.30, 0.40, 0.60, 0.90, 0.98)
lin(y = inc , time = dap)
}
