% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist.R
\name{logist}
\alias{logist}
\title{Fit Logistic model}
\usage{
logist(y, time)
}
\arguments{
\item{y}{Proportion of the disease. Must be a value between 0 and 1.}

\item{time}{Measurement time}
}
\value{
r Growth rate

y0 Initial inoculum

time Measurement time

prev Predicted data

r2 Determination coefficient between the observed and predicted data
}
\description{
Fit Logistic model
}
\examples{
dap <- seq(1, 70, 10)
inc <- c(0.01, 0.04, 0.15, 0.31, 0.65, 0.88, 0.94)
logist(y=inc, time=dap)
}
