% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monom.R
\name{monom}
\alias{monom}
\title{Fit monomolecular model}
\usage{
monom(y, time)
}
\arguments{
\item{y}{Proportion of the disease. Must be a value between 0 and 1.}

\item{time}{Measurement time}
}
\value{
r Growth rate

y0 Initial inoculum

time Measurement time

pred Predicted data

r2 Determination coefficient between the observed and predicted data
}
\description{
Fit monomolecular model
}
\examples{
dap <- seq(10, 60, 10)
inc <- c(0.18, 0.56, 0.82, 0.91, 0.96, 0.98)
monom(y = inc, time = dap)
}
