% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvi.R
\name{mvi}
\alias{mvi}
\title{MVI}
\usage{
mvi(y, time)
}
\arguments{
\item{y}{Ammount of the disease.}

\item{time}{Measurement time}
}
\value{
mvi Mean velocity index
}
\description{
MVI
}
\examples{
day <- seq(10, 60, 10)
sev <- c(0.22, NA, 0.28, 0.36, 0.39, 0.42)
plot(sev ~ day)
mvi(y = sev, time = day)
}
